Issue 12 httr (#13)
* Began work on httr-based oauth2 authentication.
Note: The creds in this are for an app that I will either delete or highly edit soon, and they aren't enough info to actually use the app anyway.

* Implemented interactive auth.
Some of the changes are cosmetic as I sorted out how things worked. Please review the activeteam vs current_team changes carefully. I THINK I fixed a bug, but it's hard to tell. We should really add tests...

* add scope managment

* Added yaml to imports to try to get build to work.

* Styled and added scopes help

* add scopes vignette

* update scopes.yml and scopes vignette

* import purrr instead of mappers

* update mac gha to trigger on every push downstream

* Log scopes during add_team_interactive
Also changed default of scopes to slackverse, since that's the expected normal use.

* Fixed add_team_token docs.

* Fixed typo in groups permissions.

* Added functions to be used in programmatic auth, such as with a Shiny app.

* Verbose should be TRUE by default.

Co-authored-by: Jonathan Sidi <yonis@metrumrg.com>
Co-authored-by: yonicd <yonicd@gmail.com>